works_browser: extract facets into their own component
